# New factorization algorithm for polynomials over algebraic number fields