--- a/diofant/polys/solvers.py
+++ b/diofant/polys/solvers.py
@@ -23,8 +23,6 @@ def eqs_to_matrix(eqs, ring):
 
 def solve_lin_sys(eqs, ring):
     """Solve a system of linear equations."""
-    assert ring.domain.is_Field
-
     # transform from equations to matrix form
     matrix = eqs_to_matrix(eqs, ring)