--- a/docs/modules/domains.rst
+++ b/docs/modules/domains.rst
@@ -41,6 +41,9 @@ Abstract Domains
 Concrete Domains
 ****************
 
+.. autoclass:: FiniteRing
+   :members:
+
 .. autoclass:: FiniteField
    :members: