--- a/docs/release/notes-0.12.rst
+++ b/docs/release/notes-0.12.rst
@@ -9,6 +9,7 @@ New features
 
 * Support modular exponentiation of :class:`~diofant.polys.rings.PolyElement`'s, see :pull:`1032`.
 * :func:`~diofant.solvers.inequalities.reduce_inequalities` support solving linear inequalities with Fourier-Motzkin elimination algorithm, see :pull:`1063`.
+* Added :class:`~diofant.domains.FiniteRing` for modular integers, see :pull:`876`.
 
 Major changes
 =============