This is a continuation of work in sympy/sympy#2468 and https://github.com/diofant/diofant/pull/712.

Todo:
- [x] ResidueRing -> FiniteRing or QuotientRing?
- [x] full test coverage
- [x] don't use `ring.dmp_LC`
- [x] split out https://github.com/sympy/sympy/commit/b5986e84801a305fafc51827ca40ef010d322925 and https://github.com/sympy/sympy/commit/f52f756b911c529c41c303924542ebf8d0f98f4c?
- [ ] better file/functions names?
- [x] changes for rref/solve_lin_sys?
- [x] rebase & cleanup
- [x] release notes
- [x] up-to-date benchmark
- [x] change this to be a default factorization method for algebraic fields?
- [ ] address TODO
- [ ] check tests

@katjasophie, I assume, I can use your unmerged yet work from sympy/sympy#2468 to the Diofant on terms of the usual SymPy license.  Please reply, if you are disagreed.

Some benchmarks on my system:
```
In [2]: R, x, y = ring('x, y', QQ.algebraic_field(sqrt(2)))
   ...: f = (x + y)*(x**2 + sqrt(2)*y)
   ...: %timeit efactor(f)
   ...: %timeit f.factor_list()
350 ms ± 3.83 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
226 ms ± 875 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [3]: R, x, y, z = ring('x, y, z', QQ.algebraic_field(root(2, 5)))
   ...: f = (x - y)*(x**2*y + root(2, 5)*y**2*z)
   ...: %timeit efactor(f)
   ...: %timeit f.factor_list()
514 ms ± 12.6 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
4.44 s ± 545 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [4]: R, x, y, z = ring('x, y, z', QQ.algebraic_field(sqrt(2), root(3, 3)))
   ...: f = (x**2 + 2*y + sqrt(2))*(root(3, 3)*x - z)
   ...: %timeit efactor(f)
   ...: %timeit f.factor_list()
2.66 s ± 141 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
47.4 s ± 2.39 s per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

New algorithm is clearly better for extensions with high degrees.