@skirpichev 2020-04-03
Using the quotient rings `Z/p^nZ` seems to be a solution for wiping out the modifications in the matrices module.  This was the major issue with this stuff so far (second one - bad code coverage).

The original author (she was in SymPy's AUTHORS) seems to be unresponsive.  @asmeurer, may I ask you for permission to merge this stuff on the terms of the usual SymPy license?