@skirpichev 2020-04-03
@jksuom, it seems that Matrix.rref() can work with coefficients in `Z/nZ`, making the matrices module modifications in the sympy/sympy#2468 - unnecessary.

I'm not so sure about naming stuff.  `ResidueRing` now is for quotient rings `Z/nZ` (and I use FiniteField's in the case `n` - prime).  `FiniteRing`?