@jksuom 2020-04-03
Clearly, FiniteRing is better than FiniteField, and probably could be used, but, of course, there are other finite rings than Z/nZ.