@skirpichev 2020-04-03
> Clearly, FiniteRing is better than FiniteField

No.  FiniteField's are in the Diofant on own purpose: they are for (surprise!) finite fields (both prime and prime power orders).

But for this PR, quotient rings Z/nZ were added: c1c07003.  I was thinking about naming them.