@jksuom 2020-04-03
That is how I understood it: Z/nZ. It could be named FiniteRing (which is better than FiniteField in SymPy).