@skirpichev 2020-04-03
QuotientRing is an alternative for ResidueRing (or FiniteRing) I was thinking of.

But that sounds as a very generic construction, which would require implementing some classes for ideals, while right now I'm interested mainly in a very limited case (principal ideals, see e.g. QuotientRingElement, which was used to represent ResidueRing, FiniteField and AlgebraicField elements).