@asmeurer 2020-04-05
I can't give permission for other people's work, but I think it would be safe to assume that any pull request to SymPy is released under the SymPy license.