@skirpichev 2020-04-06
Thank you, @asmeurer, I thought the same.

BTW, it's not a clear why https://github.com/sympy/sympy/pull/2468 wasn't merged in the SymPy.  As the bad (!=100%) code coverage isn't an issue for SymPy - the only blocking thing IMHO could be matrices/solve_lin_sys modifications.  It seems, this can be solved like I did here with the "ResidueRing" (which actually does in Diofant that the FiniteField does for SymPy).