@skirpichev 2020-04-27
BTW, doesn't help with https://github.com/sympy/sympy/issues/19196