@skirpichev 2020-11-21
I feel, this is in a relatively good shape now.  Lets try to merge.

> I think it would be safe to assume that any pull request to SymPy is released under the SymPy license.

@asmeurer, maybe you could document this somehow?

It seems apparent, that the original pr https://github.com/sympy/sympy/pull/2468 is not going to be merged in the SymPy in near future (and I can't help here for obvious to you and Ondrej reasons)...