@skirpichev 2020-11-21
Current benchmarks.

On the master:
```
In [3]: A = QQ.algebraic_field(sqrt(2), root(3, 3))
      : R, x, y, z = ring('x y z', A)
      : 

In [4]: f2, f1 = x**2 + 2*y + sqrt(2), x - z/root(3, 3)
      : f = f1*f2
      : %time assert f.factor_list() == (1, [(f1, 1), (f2, 1)])
      : 
CPU times: user 5min 59s, sys: 540 ms, total: 6min
Wall time: 6min 1s
```

On this branch:
```
In [2]: f2, f1 = x**2 + 2*y + sqrt(2), x - z/root(3, 3)
      : f = f1*f2
      : %time assert f.factor_list() == (1, [(f1, 1), (f2, 1)])
CPU times: user 24 s, sys: 83.7 ms, total: 24.1 s
Wall time: 24.4 s
```

The test was taken from https://github.com/sympy/sympy/issues/19196 (with a little adaptation: division to root(3, 3) of the second polynomial).