@skirpichev 2021-02-22
> That is how I understood it: Z/nZ. It could be named FiniteRing

In the end, this seems to be a bad idea.  Already for order=4 there are 2 (commutative) rings and 1 field, which are not isomorphic to the Z/4Z.

Maybe IntegerModRing, like Sage?